Implement adapter for executor Refactored into common macro
