+++ nomos-services/data-availability/verifier/src/network/adapters/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(feature = "libp2p")]
-pub mod libp2p;
+pub mod common;
+mod executor;
+#[cfg(feature = "libp2p")]
+pub mod validator;
